requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0